Scaffold archived indicators layout
