+++ app/(protected)/projects/[slug]/archive/page.tsx
@@ -1,5 +1,5 @@
 import ArchivedOutputs from './archived-outputs';
-
+import ArchivedOutputIndicators from './archived-output-indicators';
 export default async function ProjectArchive({
   params,
 }: {
@@ -11,6 +11,8 @@ export default async function ProjectArchive({
     <div className='space-y-8'>
       <h2 className='text-xl font-semibold'>Archived outputs</h2>
       <ArchivedOutputs slug={slug as string} />
+      <h2 className='text-xl font-semibold'>Archived output indicators</h2>
+      <ArchivedOutputIndicators slug={slug as string} />
     </div>
   );
 }
